+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 ### Changed
 - Require PHP 7.1+ and Symfony 4 components
 - `RunTestsProcessEvent` (dispatched from `run` command when initializing PHPUnit processes) now contains array of environment variables instead of ProcessBuilder. Use `setEnvironmentVars()` method to change the variables passed to the process.
+- Default browser size is now defined using class constants instead of class variables. To override the default, instead of `public static $browserWidth = ...;` use `public const BROWSER_WIDTH = ...;`.
 
 ### Removed
 - `TestUtils` class which was already deprecated in 2.1.